Agregando el metodo eliminar

+++ src/app/modules/shared/components/confirm/confirm.component.ts
@@ -1,6 +1,7 @@
 import { Component, inject, OnInit } from '@angular/core';
 import { CategoryService } from '../../services/category.service';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { ProductService } from '../../services/product.service';
 
 @Component({
   selector: 'app-confirm',
@@ -12,6 +13,7 @@ export class ConfirmComponent implements OnInit  {
   private categoryService= inject(CategoryService);
   private dialogRef = inject(MatDialogRef);
   public data = inject(MAT_DIALOG_DATA);
+  private productService = inject(ProductService);
 
   ngOnInit(): void {
   }
@@ -22,12 +24,22 @@ export class ConfirmComponent implements OnInit  {
 
   delete(){
     if(this.data != null){
-    this.categoryService.deleteCategorie(this.data.id)
-    .subscribe((data:any) => {
-      this.dialogRef.close(1);
-    },(error: any) => {
-      this.dialogRef.close(2);
-    })
+
+      if(this.data.module =="category"){
+          this.categoryService.deleteCategorie(this.data.id)
+          .subscribe((data:any) => {
+            this.dialogRef.close(1);
+          },(error: any) => {
+            this.dialogRef.close(2);
+      })
+    }  else if (this.data.module == "product"){
+      this.productService.deleteProduct(this.data.id)
+          .subscribe((data:any) => {
+            this.dialogRef.close(1);
+          },(error: any) => {
+            this.dialogRef.close(2);
+      })
+    }
   }else{
     this.dialogRef.close(2);
   }
